Warn about using default implementation of user_agent()
It's helpful to provide a default implementation, but that means
consumers will always identify themselves as "hypertyper v0.1.0", and
that's probably not the behavior most consumers want, and it's not the
behavior I want. I should provide some sort of macro to easily implement
a a user_agent() function with the consumer's package name and version,
but for now, I'll just make a note in the docs and hope for the best.