Convert HTTPClient into HTTPClientFactory
A trait that only had static methods was a little bit weird, and it was
a little cumbersome to work with. There was going to be an issue
creating a proper user agent string in library crates that in turn used
this crate. A factory is much more flexible, and avoids the issue of
having to come up with a way to create a user agent string based on the
consumer's package name and version rather than this crate's package
name and version.