Copy HTTPService trait definition into repository
This was previously defined and used in my usaidwat project, but it is
generally useful across many applications, so I am refactoring it into
this reusable crate.

This is a fairly niche implementation of a get() method and is subject
to change; namely, I'd like it to return either a string or a JSON
object, depending on the context in which it is called.

Conversely, the post() method does return a JSON object, but perhaps it
should be possible to return a string, as well, for parity with the
get() method.

Well, this whole library is a work in progress at the moment.