Split HTTPService into HTTPGet and HTTPPost types
In practice, many HTTP services I implemented only needed to support GET
requests or POST requests, but not both, and I had to cheat by calling
`unimplemented()!` in methods I was not interested in supporting. But
calling `unimplemented()!` supports the nice type system I am creating
in Hypertyper. Separating support into two types makes this code more
flexible and makes better use of the type system, and it is trivial for
consumers of the code to implement HTTPService if they support both
HTTPGet and HTTPPost.

Closes #8.