Provide HTTPService implementation for unit tests
HTTPTestService can be dropped into unit tests to provide a standardized
way to mock HTTP services. It is derivative of the testing client I have
used in my usaidwat project, although not exactly the same
implementation.

This is provided because it is useful for many consumers of this
library, although at the present, it may not be complete, so as with the
rest of Hypertyper, the API of HTTPTestService is still in flux. I will
make adaptations as I see how it is actually used.

Closes #3.